xserver: stop bcopy from going really slow
The outport is most likely unnecessary on any currently used hardware,
the byte copy is necessary from what I know on IA64 and friends so leave it.

Add a new API entry point which lets a driver select the old behaviour if
such a needs is ever found.

This gives me ~20% speed up on startup on 945 hardware.