GL: Add GLX compile flags lost in modular X server changes
RISC chips that trap on unaligned loads and stores need to
define __GLX_ALIGN64.  This used to get added to the cflags
in the old *.cf files but it no longer does in the modular
X server.

Also, Alpha needs to pass -mieee to the compiler as well.

This is a simple backport of a patch that debian, and probably other
distributions, have been applying forever.  To the best of my
knowledge the patch was written by Jurij Smakov.  See Debian bug
number #388125.

I just checked and this has been rotting for more than a year in
freedesktop bugzilla as #8392.

Signed-off-by: David S. Miller <davem@davemloft.net>