glx: Use glapi sources from the mesa tree
With recent mesa HEAD, the glapi sources used only in the xserver glx
module are carried in the mesa tree. Previously, these were generated
separately and committed to the xserver tree.

The build is changed to symlink these files like the other mesa sources.
This reduces the chance for mismatches between mesa's glX API and the
xserver's glX API.