Revert "xfree86: plug a memory leak in xf86LoadModules."
This reverts commit 3abce3ea2b1f43bd111664d4a57e5ccd282acab0 and
6cbaf15e6109ba77819c4070f5b46c78237ec460.

The memory returned to xf86LoadModule was allocated in doLoadModule, which
calls the respective module's PreInit. As it turns out, input and output
drivers store a pointer to the module elswhere, so freeing it in
xf86LoadModule is a bad idea.

For further reference: hw/xfree86/common/xf86Helper.c
    Input drivers: xf86InputDriverList[blah]->module = module;
    Output drivers: xf86DriverList[blah]->module = module;

Unloading the module would not look pretty then.